Bug 1208536: Run sync_projects in parallel.
Also removes the deleted field from Translations and returns
to deleting Translations directly. Now that we track changes
via ChangedEntityLocale, it is no longer needed, and removing
it allows us to parallelize sync_projects without coordinating
to remove deleted translations once we're done.
