+++ pontoon/sync/management/commands/sync_projects.py
@@ -1,26 +1,7 @@
-from collections import Counter
-from datetime import datetime
-
-from django.contrib.auth.models import User
 from django.core.management.base import BaseCommand, CommandError
-from django.db import connection, transaction
-from django.template.loader import render_to_string
-from django.utils import timezone
 
-from bulk_update.helper import bulk_update
-
-from pontoon.administration.vcs import CommitToRepositoryException
-from pontoon.base.models import (
-    ChangedEntityLocale,
-    Entity,
-    Locale,
-    Project,
-    Resource,
-    Translation,
-    update_stats
-)
-from pontoon.base.utils import match_attr
-from pontoon.sync.vcs_models import VCSProject
+from pontoon.base.models import Project
+from pontoon.sync.tasks import sync_project
 
 
 class Command(BaseCommand):
@@ -44,27 +25,11 @@ class Command(BaseCommand):
             help='Do not pull new commits from VCS'
         )
 
-    def log(self, msg, *args, **kwargs):
-        """Log a message to the console."""
-        self.stdout.write(msg.format(*args, **kwargs))
-
-    def info(self, msg, *args, **kwargs):
-        """Log a message to the console if --verbosity=1 or more."""
-        if self.verbosity >= 1:
-            self.log(msg, *args, **kwargs)
-
-    def debug(self, msg, *args, **kwargs):
-        """Log a message to the console if --verbosity=2."""
-        if self.verbosity == 2:
-            self.log(msg, *args, **kwargs)
-
     def handle(self, *args, **options):
-        self.verbosity = options['verbosity']
-        self.no_commit = options['no_commit']
-        self.no_pull = options['no_pull']
-
-        self.log('SYNC PROJECTS: start')
-
+        """
+        Collect the projects we want to sync and trigger worker jobs to
+        sync each one.
+        """
         projects = Project.objects.filter(disabled=False)
         if args:
             projects = projects.filter(slug__in=args)
@@ -74,428 +39,12 @@ class Command(BaseCommand):
 
         for project in projects:
             if not project.can_commit:
-                self.log(u'Skipping project {0}, cannot commit to repository.'
-                         .format(project.name))
+                self.stdout.write(u'Skipping project {0}, cannot commit to repository.'
+                                  .format(project.name))
             else:
-                self.handle_project(project)
-        self.log('SYNC PROJECTS: done')
-
-        # Once we've synced, we can delete all translations scheduled
-        # for deletion.
-        Translation.deleted_objects.all().delete()
-
-    @transaction.atomic
-    def handle_project(self, db_project):
-        # Mark "now" at the start of sync to avoid messing with
-        # translations submitted during sync.
-        now = timezone.now()
-
-        # Pull changes from VCS and update what we know about the files.
-        if not self.no_pull:
-            repos_changed = self.pull_changes(db_project)
-        else:
-            repos_changed = True  # Assume changed.
-
-        # If the repos haven't changed since the last sync and there are
-        # no Pontoon-side changes for this project, quit early.
-        db_changed = (ChangedEntityLocale.objects
-                      .filter(entity__resource__project=db_project)
-                      .exists())
-        if not repos_changed and not db_changed:
-            self.log('Skipping project {0}, no changes detected.'.format(db_project.slug))
-            return
-
-        vcs_project = VCSProject(db_project)
-        self.update_resources(db_project, vcs_project)
-
-        # Collect all entities across VCS and the database and get their
-        # keys so we can match up matching entities.
-        vcs_entities = self.get_vcs_entities(vcs_project)
-        db_entities = self.get_db_entities(db_project)
-        entity_keys = set().union(db_entities.keys(), vcs_entities.keys())
-
-        changeset = ChangeSet(db_project, vcs_project, now)
-        for key in entity_keys:
-            db_entity = db_entities.get(key, None)
-            vcs_entity = vcs_entities.get(key, None)
-            self.handle_entity(changeset, db_project, key, db_entity, vcs_entity)
-
-        # Apply the changeset to the files, commit them, and update stats
-        # entries in the DB.
-        changeset.execute()
-        if not self.no_commit:
-            self.commit_changes(db_project, vcs_project, changeset)
-        self.update_stats(db_project, vcs_project, changeset)
-
-        # Clear out the list of changed locales for entity in this
-        # project now that we've finished syncing.
-        (ChangedEntityLocale.objects
-            .filter(entity__resource__project=db_project, when__lte=now)
-            .delete())
-
-        # Clean up any duplicate approvals at the end of sync right
-        # before we commit the transaction to avoid race conditions.
-        with connection.cursor() as cursor:
-            cursor.execute("""
-                UPDATE base_translation AS b
-                SET approved = FALSE, approved_date = NULL
-                WHERE deleted IS NULL
-                  AND approved_date !=
-                    (SELECT max(approved_date)
-                     FROM base_translation
-                     WHERE entity_id = b.entity_id
-                       AND deleted IS NULL
-                       AND locale_id = b.locale_id
-                       AND (plural_form = b.plural_form OR plural_form IS NULL));
-            """)
-
-        self.log(u'Synced project {0}', db_project.slug)
-
-    def handle_entity(self, changeset, db_project, key, db_entity, vcs_entity):
-        """
-        Determine what needs to be synced between the database and VCS versions
-        of a single entity and log what needs to be changed in the changeset.
-        """
-        if vcs_entity is None:
-            if db_entity is None:
-                # This should never happen. What? Hard abort.
-                raise CommandError('No entities found for key {0}'.format(key))
-            else:
-                # VCS no longer has the entity, remove it from Pontoon.
-                changeset.obsolete_db_entity(db_entity)
-        elif db_entity is None:
-            # New VCS entities are added to Pontoon.
-            changeset.create_db_entity(vcs_entity)
-        else:
-            for locale in db_project.locales.all():
-                if not vcs_entity.has_translation_for(locale.code):
-                    # VCS lacks an entity for this locale, so we can't
-                    # pull updates nor edit it. Skip it!
-                    continue
-
-                if db_entity.has_changed(locale):
-                    # Pontoon changes overwrite whatever VCS has.
-                    changeset.update_vcs_entity(locale.code, db_entity, vcs_entity)
-
-                else:
-                    # If Pontoon has nothing or has not changed, and the VCS
-                    # still has the entity, update Pontoon with whatever may
-                    # have changed.
-                    changeset.update_db_entity(locale.code, db_entity, vcs_entity)
-
-    def update_resources(self, db_project, vcs_project):
-        """Update the database on what resource files exist in VCS."""
-        relative_paths = vcs_project.resources.keys()
-        db_project.resource_set.exclude(path__in=relative_paths).delete()
-
-        for relative_path, vcs_resource in vcs_project.resources.items():
-            resource, created = db_project.resource_set.get_or_create(path=relative_path)
-            resource.format = Resource.get_path_format(relative_path)
-            resource.entity_count = len(vcs_resource.entities)
-            resource.save()
-
-    def update_stats(self, db_project, vcs_project, changeset):
-        """Update the Stats entries in the database."""
-        for resource in db_project.resource_set.all():
-            for locale in db_project.locales.all():
-                # We only want to create/update the stats object if the resource
-                # exists in the current locale, UNLESS the file is asymmetric.
-                vcs_resource = vcs_project.resources[resource.path]
-                resource_exists = vcs_resource.files.get(locale) is not None
-                if resource_exists or resource.is_asymmetric:
-                    update_stats(resource, locale)
-
-    def get_vcs_entities(self, vcs_project):
-        return {self.entity_key(entity): entity for entity in vcs_project.entities}
-
-    def get_db_entities(self, db_project):
-        entities = (Entity.objects
-                    .select_related('resource')
-                    .prefetch_related('changed_locales')
-                    .filter(resource__project=db_project, obsolete=False))
-        return {self.entity_key(entity): entity for entity in entities}
-
-    def entity_key(self, entity):
-        """
-        Generate a key for the given entity that is unique within the
-        project.
-        """
-        key = entity.key or entity.string
-        return ':'.join([entity.resource.path, key])
-
-    def pull_changes(self, db_project):
-        """
-        Update the local files with changes from the VCS. Returns
-        whether any of the updated repos have changed since the last
-        sync, based on the revision numbers.
-        """
-        changed = False
-        for repo in db_project.repositories.all():
-            repo_revisions = repo.pull()
-
-            # If any revision is None, we can't be sure if a change
-            # happened or not, so we default to assuming it did.
-            unsure_change = None in repo_revisions.values()
-            if unsure_change or repo_revisions != repo.last_synced_revisions:
-                changed = True
-
-            repo.last_synced_revisions = repo_revisions
-            repo.save()
-
-        return changed
-
-    def commit_changes(self, db_project, vcs_project, changeset):
-        """Commit the changes we've made back to the VCS."""
-        for locale in db_project.locales.all():
-            authors = changeset.commit_authors_per_locale.get(locale.code, [])
-
-            # Use the top translator for this batch as commit author, or
-            # the fake Pontoon user if there are no authors.
-            if len(authors) > 0:
-                commit_author = Counter(authors).most_common(1)[0][0]
-            else:
-                commit_author = User(first_name="Pontoon", email="pontoon@mozilla.com")
-
-            commit_message = render_to_string('commit_message.jinja', {
-                'locale': locale,
-                'project': db_project,
-                'authors': authors
-            })
-
-            locale_path = vcs_project.locale_directory_path(locale.code)
-            try:
-                repo = db_project.repository_for_path(locale_path)
-                result = repo.commit(commit_message, commit_author, locale_path)
-            except (CommitToRepositoryException, ValueError) as err:
-                result = {'message': unicode(err)}
-
-            if result is not None:
-                self.log(
-                    u'Committing project {project.name} for {locale.name} '
-                    u'({locale.code}) failed: {reason}',
-                    project=db_project,
-                    locale=locale,
-                    reason=result['message']
+                self.stdout.write(u'Scheduling sync for project {0}.'.format(project.name))
+                sync_project.delay(
+                    project.pk,
+                    no_pull=options['no_pull'],
+                    no_commit=options['no_commit']
                 )
-
-
-class ChangeSet(object):
-    """
-    Stores a set of changes to be made to the database and the
-    translations stored in VCS. Once all the necessary changes have been
-    stored, execute all the changes at once efficiently.
-    """
-    def __init__(self, db_project, vcs_project, now):
-        """
-        :param now:
-            Datetime to use for marking when approvals happened.
-        """
-        self.db_project = db_project
-        self.vcs_project = vcs_project
-        self.now = now
-
-        self.executed = False
-        self.changes = {
-            'update_vcs': [],
-            'update_db': [],
-            'obsolete_db': [],
-            'create_db': []
-        }
-
-        self.entities_to_update = []
-        self.translations_to_update = []
-        self.translations_to_create = []
-
-        self.commit_authors_per_locale = {}
-
-    def update_vcs_entity(self, locale_code, db_entity, vcs_entity):
-        """
-        Replace the translations in VCS with the translations from the
-        database.
-        """
-        self.changes['update_vcs'].append((locale_code, db_entity, vcs_entity))
-
-    def create_db_entity(self, vcs_entity):
-        """Create a new entity in the database."""
-        self.changes['create_db'].append(vcs_entity)
-
-    def update_db_entity(self, locale_code, db_entity, vcs_entity):
-        """Update the database with translations from VCS."""
-        self.changes['update_db'].append((locale_code, db_entity, vcs_entity))
-
-    def obsolete_db_entity(self, db_entity):
-        """Mark the given entity as obsolete."""
-        self.changes['obsolete_db'].append(db_entity.pk)
-
-    def execute(self):
-        """
-        Execute the changes stored in this changeset. Execute can only
-        be called once per changeset; subsequent calls raise a
-        RuntimeError, even if the changes failed.
-        """
-        if self.executed:
-            raise RuntimeError('execute() can only be called once per changeset.')
-        else:
-            self.executed = True
-
-        # Store locales and resources for FK relationships.
-        self.locales = {l.code: l for l in Locale.objects.all()}
-        self.resources = {r.path: r for r in self.db_project.resource_set.all()}
-
-        # Perform the changes and fill the lists for bulk creation and
-        # updating.
-        self.execute_update_vcs()
-        self.execute_create_db()
-        self.execute_update_db()
-        self.execute_obsolete_db()
-
-        # Apply the built-up changes to the DB
-        if len(self.entities_to_update) > 0:
-            bulk_update(self.entities_to_update, update_fields=[
-                'resource',
-                'string',
-                'string_plural',
-                'key',
-                'comment',
-                'order',
-                'source'
-            ])
-
-        Translation.objects.bulk_create(self.translations_to_create)
-        if len(self.translations_to_update) > 0:
-            bulk_update(self.translations_to_update, update_fields=[
-                'entity',
-                'locale',
-                'string',
-                'plural_form',
-                'approved',
-                'approved_user_id',
-                'approved_date',
-                'fuzzy',
-                'extra'
-            ])
-
-    def execute_update_vcs(self):
-        resources = self.vcs_project.resources
-        changed_resources = set()
-
-        for locale_code, db_entity, vcs_entity in self.changes['update_vcs']:
-            changed_resources.add(resources[db_entity.resource.path])
-
-            vcs_translation = vcs_entity.translations[locale_code]
-            db_translations = (db_entity.translation_set
-                               .filter(approved=True, locale__code=locale_code))
-
-            # If no DB translations are fuzzy, set fuzzy to False.
-            # Otherwise, it's true.
-            vcs_translation.fuzzy = any(t for t in db_translations if t.fuzzy)
-
-            if len(db_translations) > 0:
-                last_translation = max(
-                    db_translations,
-                    key=lambda t: t.date or timezone.make_aware(datetime.min)
-                )
-                vcs_translation.last_updated = last_translation.date
-                vcs_translation.last_translator = last_translation.user
-
-            # Replace existing translations with ones from the database.
-            vcs_translation.strings = {
-                db.plural_form: db.string for db in db_translations
-            }
-
-            # Track which translators were involved.
-            self.commit_authors_per_locale[locale_code] = [t.user for t in db_translations if t.user]
-
-        for resource in changed_resources:
-            resource.save()
-
-    def get_entity_updates(self, vcs_entity):
-        """
-        Return a dict of the properties and values necessary to create
-        or update a database entity from a VCS entity.
-        """
-        return {
-            'resource': self.resources[vcs_entity.resource.path],
-            'string': vcs_entity.string,
-            'string_plural': vcs_entity.string_plural,
-            'key': vcs_entity.key,
-            'comment': '\n'.join(vcs_entity.comments),
-            'order': vcs_entity.order,
-            'source': vcs_entity.source
-        }
-
-    def execute_create_db(self):
-        for vcs_entity in self.changes['create_db']:
-            entity = Entity(**self.get_entity_updates(vcs_entity))
-            entity.save()  # We can't use bulk_create since we need a PK
-
-            for locale_code, vcs_translation in vcs_entity.translations.items():
-                for plural_form, string in vcs_translation.strings.items():
-                    self.translations_to_create.append(Translation(
-                        entity=entity,
-                        locale=self.locales[locale_code],
-                        string=string,
-                        plural_form=plural_form,
-                        approved=not vcs_translation.fuzzy,
-                        approved_date=self.now if not vcs_translation.fuzzy else None,
-                        fuzzy=vcs_translation.fuzzy
-                    ))
-
-    def execute_update_db(self):
-        for locale_code, db_entity, vcs_entity in self.changes['update_db']:
-            for field, value in self.get_entity_updates(vcs_entity).items():
-                setattr(db_entity, field, value)
-
-            if db_entity.is_dirty(check_relationship=True):
-                self.entities_to_update.append(db_entity)
-
-            # Update translations for the entity.
-            vcs_translation = vcs_entity.translations[locale_code]
-            db_translations = db_entity.translation_set.filter(
-                locale__code=locale_code,
-            )
-            approved_translations = []
-
-            for plural_form, string in vcs_translation.strings.items():
-                # Check if we need to modify an existing translation or
-                # create a new one.
-                db_translation = match_attr(db_translations,
-                                            plural_form=plural_form,
-                                            string=string)
-                if db_translation:
-                    if not db_translation.approved:
-                        db_translation.approved = True
-                        db_translation.approved_date = self.now
-                    db_translation.fuzzy = vcs_translation.fuzzy
-                    db_translation.extra = vcs_translation.extra
-
-                    if db_translation.is_dirty():
-                        self.translations_to_update.append(db_translation)
-                    if not db_translation.fuzzy:
-                        approved_translations.append(db_translation)
-                else:
-                    self.translations_to_create.append(Translation(
-                        entity=db_entity,
-                        locale=self.locales[locale_code],
-                        string=string,
-                        plural_form=plural_form,
-                        approved=not vcs_translation.fuzzy,
-                        approved_date=self.now if not vcs_translation.fuzzy else None,
-                        fuzzy=vcs_translation.fuzzy,
-                        extra=vcs_translation.extra
-                    ))
-
-            # Any existing translations that were not approved get unapproved.
-            for translation in db_translations.filter(approved_date__lte=self.now):
-                if translation not in approved_translations:
-                    translation.approved = False
-                    translation.approved_user = None
-                    translation.approved_date = None
-
-                    if translation.is_dirty():
-                        self.translations_to_update.append(translation)
-
-    def execute_obsolete_db(self):
-        (Entity.objects
-            .filter(pk__in=self.changes['obsolete_db'])
-            .update(obsolete=True))
